requirements.txt
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.3
django-taggit==3.0.0
markdown==3.4.1
psycopg2-binary==2.9.3